# Backend/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: FirstRegistrars BuildWise API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /tasks/{id}:
    get:
      operationId: getTask
      tags: [tasks]
      summary: Get a task
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
    put:
      operationId: updateTask
      tags: [tasks]
      summary: Update a task
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTaskBody"
      responses:
        "200":
          description: Updated task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
    delete:
      operationId: deleteTask
      tags: [tasks]
      summary: Delete a task
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ---- SPRINTS ----
...
  /sprints/{id}:
    put:
      operationId: updateSprint
      tags: [sprints]
      summary: Update a sprint
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateSprintBody"
      responses:
        "200":
          description: Updated sprint
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Sprint"
    delete:
      operationId: deleteSprint
      tags: [sprints]
      summary: Delete a sprint
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ---- COMMENTS ----
  /tasks/{taskId}/comments:
    get:
      operationId: listComments
      tags: [comments]
      summary: List comments on a task
      parameters:
        - name: taskId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Comments list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Comment"
    post:
      operationId: createComment
      tags: [comments]
      summary: Create a comment
      parameters:
        - name: taskId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCommentBody"
      responses:
        "201":
          description: Created comment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Comment"

  # ---- VENDORS ----
...
components:
  schemas:
...
    Task:
      type: object
      properties:
        id:
          type: integer
        projectId:
          type: integer
        sprintId:
          type: integer
          nullable: true
        title:
          type: string
        description:
          type: string
          nullable: true
        status:
          type: string
          enum: [backlog, todo, in_progress, in_review, done]
        priority:
          type: string
          enum: [low, medium, high, critical]
        type:
          type: string
          enum: [story, task, bug, epic, subtask]
        assigneeId:
          type: integer
          nullable: true
        reporterId:
          type: integer
          nullable: true
        storyPoints:
          type: integer
          nullable: true
        dueDate:
          type: string
          format: date
          nullable: true
        label:
          type: string
          nullable: true
        position:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - projectId
        - title
        - status
        - priority
        - type
        - position
        - createdAt
        - updatedAt
...
    UpdateTaskBody:
      type: object
      properties:
        sprintId:
          type: integer
          nullable: true
        title:
          type: string
        description:
          type: string
          nullable: true
        status:
          type: string
          enum: [backlog, todo, in_progress, in_review, done]
        priority:
          type: string
          enum: [low, medium, high, critical]
        type:
          type: string
          enum: [story, task, bug, epic, subtask]
        assigneeId:
          type: integer
          nullable: true
        reporterId:
          type: integer
          nullable: true
        storyPoints:
          type: integer
          nullable: true
        dueDate:
          type: string
          format: date
          nullable: true
        label:
          type: string
          nullable: true
        position:
          type: integer

    # SPRINTS
    Sprint:
      type: object
      properties:
        id:
          type: integer
        projectId:
          type: integer
        name:
          type: string
        goal:
          type: string
          nullable: true
        status:
          type: string
          enum: [planned, active, completed]
        startDate:
          type: string
          format: date
          nullable: true
        endDate:
          type: string
          format: date
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - projectId
        - name
        - status
        - createdAt
...
    UpdateSprintBody:
      type: object
      properties:
        name:
          type: string
        goal:
          type: string
          nullable: true
        status:
          type: string
          enum: [planned, active, completed]
        startDate:
          type: string
          format: date
          nullable: true
        endDate:
          type: string
          format: date
          nullable: true

    # COMMENTS
    Comment:
      type: object
      properties:
        id:
          type: integer
        taskId:
          type: integer
        authorId:
          type: integer
          nullable: true
        content:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - taskId
        - content
        - createdAt
        - updatedAt
    CreateCommentBody:
      type: object
      properties:
        authorId:
          type: integer
          nullable: true
        content:
          type: string
      required:
        - content

    # VENDORS
...